+++ R/run_schedule.R
@@ -12,6 +12,12 @@
 #' whether it is scheduled to run at the current time using some simple time arithmetic. We assume
 #' `run_schedule(schedule, check_datetime = Sys.time())`, but this need not be the case.
 #'
+#' ## Output
+#'
+#' `run_schedule()` returns a list with two elements: `status` and `artifacts`. Status is a data.frame where
+#' each row is a pipeline and the columns are information about the pipeline status, execution time, etc. Artifacts
+#' are any values returned from pipelines.
+#'
 #' ## Pipelines with arguments (resources)
 #'
 #' If a pipeline takes an argument that doesn't include a default value, these can be supplied
@@ -21,6 +27,7 @@
 #' the same argument but requiring different values.
 #'
 #' ## Running in parallel
+#'
 #' Pipelines can be run in parallel using the `cores` argument. First, you must run `future::plan(future::multisession)`
 #' in the orchestrator. Then, supply the desired number of cores to the `cores` argument. Note that
 #' console output appears different in multicore mode.
@@ -39,7 +46,7 @@
 #' @param log_file_max_bytes numeric specifying the maximum number of bytes allowed in the log file before purging the log (within a margin of error)
 #' @param quiet silence metrics to the console (default = `FALSE`)
 #'
-#' @return data.frame of pipeline statuses
+#' @return list with named elements `status` and `artifacts`
 #' @importFrom R.utils countLines
 #' @export
 #' @examples
@@ -184,6 +191,11 @@ run_schedule <- function(
 
     tictoc::tic(quiet = quiet)
 
+    # Create external vars for these (this won't work in parallel)
+    orchestrator_context <- new.env()
+    assign("start_times", lubridate::POSIXct(tz = "UTC"), envir = orchestrator_context)
+    assign("end_times", lubridate::POSIXct(tz = "UTC"), envir = orchestrator_context)
+
     # Execute the schedule (possibly in parallel)
     runs <- mapper_fun(
       list(
@@ -197,31 +209,49 @@ run_schedule <- function(
         ~{
           if (..3 || !..5) {
             cli::cli_alert("{cli::col_silver(..1)} { ..2}")
+            assign(
+              "start_times",
+              c(get("start_times", envir = orchestrator_context), NA),
+              envir = orchestrator_context
+            )
+            assign(
+              "end_times",
+              c(get("end_times", envir = orchestrator_context), NA),
+              envir = orchestrator_context
+            )
             return(
               list(
                 warnings = NULL,
                 messages = NULL,
-                skip = TRUE,
-                start_time = NA,
-                end_time = NA
+                skip = TRUE
               )
             )
           } else {
             cli::cli_progress_step("{cli::col_silver(..1)} {.pkg { ..2}}")
-            run_schedule_entry(
-              ..1,
-              ..2,
-              resources = resources,
-              log_file = log_file,
-              log_level = ..4,
-              log_file_max_bytes = log_file_max_bytes
+            assign(
+              "start_times",
+              c(get("start_times", envir = orchestrator_context), lubridate::now("UTC")),
+              envir = orchestrator_context
             )
+            tryCatch({
+              run_schedule_entry(
+                ..1,
+                ..2,
+                resources = resources,
+                log_file = log_file,
+                log_level = ..4,
+                log_file_max_bytes = log_file_max_bytes
+              )
+            }, finally = {
+              assign(
+                "end_times",
+                c(get("end_times", envir = orchestrator_context),
+                  lubridate::now("UTC")),
+                envir = orchestrator_context
+                )
+            })
           }
         },
-        otherwise = list(
-          start_time = NA,
-          end_time = NA
-        ),
         quiet = TRUE
       )
     ) |>
@@ -231,23 +261,34 @@ run_schedule <- function(
 
     elapsed <- tictoc::toc(quiet = TRUE)
 
-    # Get the results as a tibble
-    results <- purrr::imap(
+    # Get the start/end times from the orchestrator context
+    start_times <- get("start_times", envir = orchestrator_context)
+    end_times <- get("end_times", envir = orchestrator_context)
+    start_times <- if (length(start_times) == 0) NA else start_times
+    end_times <- if (length(end_times) == 0) NA else end_times
+
+    # Get the status as a tibble
+    status <- purrr::imap(
       runs,
       ~dplyr::tibble(
         pipe_id = as.integer(.y),
-        errors = length(.x$error),
+        errors = as.integer(length(.x$error) > 0),
         warnings = length(.x$result$warnings),
-        messages = length(.x$result$messages),
-        pipeline_started = .x$result$start_time,
-        pipeline_ended = .x$result$end_time
+        messages = length(.x$result$messages)
       )
     ) |>
-      purrr::list_rbind()
+      purrr::list_rbind() |>
+      dplyr::mutate(
+        pipeline_started = start_times,
+        pipeline_ended = end_times
+      )
+
+    # Modify the names
+    runs <- runs |>
+      purrr::set_names(glue::glue("{schedule$pipe_name} {schedule$script_path}"))
 
     # Get the errors
     run_errors <- runs |>
-      purrr::set_names(glue::glue("{schedule$pipe_name} {schedule$script_path}")) |>
       purrr::map(
         ~.x$error
       ) |>
@@ -255,7 +296,6 @@ run_schedule <- function(
 
     # Get the warnings
     run_warnings <- runs |>
-      purrr::set_names(glue::glue("{schedule$pipe_name} {schedule$script_path}")) |>
       purrr::map(
         ~.x$result$warnings
       ) |>
@@ -263,7 +303,6 @@ run_schedule <- function(
 
     # Get the messages
     run_messages <- runs |>
-      purrr::set_names(glue::glue("{schedule$pipe_name} {schedule$script_path}")) |>
       purrr::map(
         ~.x$result$messages
       ) |>
@@ -276,7 +315,7 @@ run_schedule <- function(
 
     # Create status table
     status_table <- schedule |>
-      dplyr::left_join(results, by = "pipe_id") |>
+      dplyr::left_join(status, by = "pipe_id") |>
       dplyr::mutate(
         success = ifelse(invoked, errors == 0, NA),
         dplyr::across(
@@ -297,11 +336,19 @@ run_schedule <- function(
         next_run
       )
 
+    # Get the artifacts as a named list
+    artifacts <- runs |>
+      purrr::map(
+        ~.x$result$artifacts
+      ) |>
+      purrr::discard(is.null)
+
     # Get the number of statuses
-    total <- sum(status_table$invoked)
+    total <- nrow(status_table)
+    invoked <- sum(status_table$invoked)
     error_count <- length(run_errors)
     skip_count <- sum(!status_table$invoked)
-    success_count <- total - error_count
+    success_count <- invoked - error_count
     warning_count <- length(run_warnings)
 
     cli::cli_h3("Pipeline execution completed {cli::col_silver(cli::symbol$stop)} | {elapsed$callback_msg}")
@@ -344,16 +391,21 @@ run_schedule <- function(
       cli::cli_ul(next_run_strs)
     }
 
-    return(status_table)
+    output <- list(
+      status = status_table,
+      artifacts = artifacts
+    )
+
+    return(output)
   }
 
   if (quiet) {
     run_schedule_fun <- purrr::quietly(run_schedule_fun)
-    status_table <- run_schedule_fun() |>
+    output <- run_schedule_fun() |>
       purrr::pluck("result")
   } else {
-    status_table <- run_schedule_fun()
+    output <- run_schedule_fun()
   }
 
-  return(status_table)
+  return(output)
 }